% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmbeR-package.R
\docType{package}
\name{bmbeR-package}
\alias{bmbeR}
\alias{bmbeR-package}
\title{bmbeR: Prior-Focused Bayesian Model Building and Evaluation}
\description{
bmbeR wraps a principled Bayesian workflow (Gelman et al., 2020) around
regression models fitted with \code{\link[rstanarm:stan_glm]{rstanarm::stan_glm()}}, with an emphasis on
the part of the workflow that existing packages leave to the analyst: the
prior.

The functions map onto the stages of the workflow:\tabular{ll}{
   Stage \tab Functions \cr
   Specify priors \tab \code{\link[=prior_spec]{prior_spec()}}, \code{\link[=prior_config]{prior_config()}}, \code{\link[=empirical_bayes_priors]{empirical_bayes_priors()}} \cr
   Check priors before fitting \tab \code{\link[=prior_predictive_check]{prior_predictive_check()}} \cr
   Fit and check computation \tab \code{\link[=fit_model_with_prior]{fit_model_with_prior()}}, \code{\link[=check_convergence]{check_convergence()}} \cr
   Quantify prior influence \tab \code{\link[=prior_sensitivity]{prior_sensitivity()}}, \code{\link[=sensitivity_analysis]{sensitivity_analysis()}}, \code{\link[=plot_prior_posterior]{plot_prior_posterior()}} \cr
   Visualise \tab \code{\link[=generate_plot]{generate_plot()}}, \code{\link[=plot_posterior_distributions]{plot_posterior_distributions()}} \cr
   Evaluate predictions \tab \code{\link[=evaluate_model_performance]{evaluate_model_performance()}} \cr
   Report \tab \code{\link[=workflow_report]{workflow_report()}} \cr
}
}
\section{Global options}{

bmbeR never changes global state (options, the random seed, or the ggplot2
theme). To run chains in parallel, set \code{options(mc.cores = 4)} (or
\code{parallel::detectCores()}) yourself before fitting.
}

\references{
Gelman, A., Vehtari, A., Simpson, D., Margossian, C. C., Carpenter, B.,
Yao, Y., Kennedy, L., Gabry, J., Bürkner, P.-C., & Modrák, M. (2020).
Bayesian workflow. \emph{arXiv:2011.01808}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elkronos.github.io/bmbeR/}
  \item \url{https://github.com/elkronos/bmbeR}
  \item Report bugs at \url{https://github.com/elkronos/bmbeR/issues}
}

}
\author{
\strong{Maintainer}: elkronos \email{jchase.msu@gmail.com}

}
\keyword{internal}
